Add an "external ID" to all indexing operations logs.
When triggering indexing operation from the outside asynchronously, one might
want to keep track of whether its actions has not started yet, is underway, or
has completed.

This commit adds a notion of "external ID", stored both in the indexing lock
and in the audit log, that can optionally be provided when triggering an
indexing operation.